refactor: 스터디 진행 전 관련 플로우 컴포넌트 전면 수정
